+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Fixed
 ### Changed
 - FF-1429 - Updated SonarAnalyzer.CSharp to 8.27.0.35380
+- FF-1429 - Updated Microsoft.CodeAnalysis.CSharp.Workspaces to 3.11.0
 ### Removed
 ### Deployment Changes