+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Fixed
 ### Changed
 - FF-1429 - Updated SonarAnalyzer.CSharp to 8.28.0.36354
+- FF-1429 - Updated SonarAnalyzer.CSharp to 8.29.0.36737
 ### Removed
 ### Deployment Changes